Fix up reference to the apparmor profile
The old way was to read the profile in from the config
and if it didn't exist to use a default profile. We will
continue with the default profile eventually, but for
now I just want the profile to be a part of the language
object that we access directly. It's more readable