Formatting and parsing of the new return value for run
Instead of returning a long tuple we return an object
instance that holds all the data about the code execution
We now need to parse that data out for the json return

Notice also I changed 'exitstatus' to 'returncode'. That
is to match what we have in our current code base. I think
'exitstatus' is probably a better name though