Switch to more verbose output
The old way of doing things didn't have enough data. For
instance, it did not clearly delineate when the program
had been successfully executed but had returned a nonzero
returncode. The new way indicates when the program
successfully ran to completion and its outputs

It also now has status messages closer to what we used
in CodeVue's original backend