Fix up the language tests to work with the new output format
This involved changing the fact that they store 'error', sometimes
as None, and instead prefer specifying a status that defaults
to success. It also involved changing the way we pattern match
on output to check the first run's output.

Tcl was a special case and required a change to the apparmor
profile and removing a test. I believe I failed to run the full
suite of tests before as I was getting syntax errors in the test
itself. My basic understanding of tcl right now indicates that
the fork() command is not installed by default, and I don't think
it is worth it to get it installed.