Fix the objective-c profile
There were several things broken. I wasn't using the correct comilation flags
provided by gnustep-config --objc-flags. Also the apparmor profile didn't allow
for access to a few things that objective-c needs to run, like the /proc/
portion of the filesystem.

Ultimately we worked all of this out with a great deal of help from Pete Boden
who came up with the source for the new language test I've added that shows
that we can actually execute Objective-C code that reads stdin and prints to
stdout, which we didn't have before