Add a very brief sleep to allow subprocess to start
Without this we can notify the conditional so fast that
the new thread hasn't yet started waiting on the conditional.
This tiny sleep helps to avoid that by delaying a bit.

It would be much better to not need this, but I can't think
of an easy way to fix it right now