Bring down the read timeout
This is because our workflows want a heartbeat every 60 seconds
That means that we need nginx to error out before that time, if
it is going to error out. That means that we need straitjacket
to try to get back to nginx in time. All of that means we need
to give a little bit of padding at each level